move runReasoning + types to the MCP server
